# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/students": {
            "get": {
                "tags": [
                    "Students"
                ],
                "summary": "Get list of students",
                "operationId": "75378cfd9c3d128e0846de3d32b101ec",
                "parameters": [
                    {
                        "name": "userType",
                        "in": "header",
                        "description": "Type of the user making the request (user, admin, manager)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "user",
                                "admin",
                                "manager"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A list of students",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Student"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/student": {
            "post": {
                "tags": [
                    "Students"
                ],
                "summary": "Create a new student",
                "operationId": "613d90c63d75c1e33836cf89dbd422e8",
                "parameters": [
                    {
                        "name": "userType",
                        "in": "header",
                        "description": "Type of the user making the request (user, admin, manager)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "user",
                                "admin",
                                "manager"
                            ]
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Student"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Student created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Student"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Student": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "example": "john.doe@example.com"
                    },
                    "age": {
                        "type": "integer",
                        "example": 25
                    },
                    "country": {
                        "type": "string",
                        "example": "USA"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Students",
            "description": "Students"
        }
    ]
}